app/pyproject.toml
[project]
name = "mcp-rest-server"
version = "0.1.0"
description = "Wrapper for MCP Clients to be called via REST on a platform"
license = "GPL-3.0"
authors = [{ name = "Matthias Kainer", email = "matthias@inxm.ai" }]
requires-python = ">=3.10"
dependencies = [
    "httpx",
    "fastapi",
    "uvicorn",
    "mcp[cli]",
    "celery",
    "PyJWT",
    "requests",
    "prometheus-fastapi-instrumentator",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "session",
    "session_manager",
    "oauth",
    "mcp_server",
]